+++ rocketgram/middlewares/middleware.py
@@ -5,7 +5,7 @@
 
 from typing import Awaitable, Union, Optional
 
-from .. import api
+from ..api import Request, Response
 
 
 class Middleware:
@@ -26,14 +26,13 @@ class Middleware:
     def process_error(self, error: Exception) -> Optional[Awaitable]:
         raise NotImplementedError
 
-    def before_request(self, request: 'api.Request') -> Union['api.Response', Awaitable['api.Request']]:
+    def before_request(self, request: Request) -> Union[Response, Awaitable[Request]]:
         raise NotImplementedError
 
-    def after_request(self, request: 'api.Request', response: 'api.Response') -> \
-            Union['api.Response', Awaitable['api.Response']]:
+    def after_request(self, request: Request, response: Response) -> Union[Response, Awaitable[Response]]:
         raise NotImplementedError
 
-    def request_error(self, request: 'api.Request', error: Exception) -> Optional[Awaitable]:
+    def request_error(self, request: Request, error: Exception) -> Optional[Awaitable]:
         raise NotImplementedError
 
 
@@ -55,12 +54,11 @@ class EmptyMiddleware(Middleware):
     def process_error(self, error: Exception) -> Optional[Awaitable]:
         pass
 
-    def before_request(self, request: 'api.Request') -> Union['api.Request', Awaitable['api.Request']]:
+    def before_request(self, request: Request) -> Union[Request, Awaitable[Request]]:
         return request
 
-    def after_request(self, request: 'api.Request', response: 'api.Response') -> \
-            Union['api.Response', Awaitable['api.Response']]:
+    def after_request(self, request: Request, response: Response) -> Union[Response, Awaitable[Response]]:
         return response
 
-    def request_error(self, request: 'api.Request', error: Exception) -> Optional[Awaitable]:
+    def request_error(self, request: Request, error: Exception) -> Optional[Awaitable]:
         pass